requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
openai
python-multipart
